Adopt all changes kindly suggested by @frant-hartm in DATAGRAPH-1008

+++ src/main/java/com/github/timtebeek/taxonomy/model/AbstractEntity.java
@@ -22,5 +22,5 @@ public abstract class AbstractEntity {
 	@Override
 	public abstract String toString();
 
-	public abstract Long getId();
+	public abstract Integer getId();
 }
