+++ src/main/java/com/github/timtebeek/taxonomy/model/Division.java
@@ -1,21 +1,22 @@
 package com.github.timtebeek.taxonomy.model;
 
-import lombok.Data;
-import lombok.EqualsAndHashCode;
 import org.neo4j.ogm.annotation.Index;
 import org.neo4j.ogm.annotation.NodeEntity;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 @NodeEntity
 @Data
 @EqualsAndHashCode(callSuper = false, of = { "divisionid" })
 public class Division extends AbstractEntity {
 	@Index(unique = true, primary = true)
-	Long	divisionid;
+	Integer	divisionid;
 
 	String	code, name, comments;
 
 	@Override
-	public Long getId() {
+	public Integer getId() {
 		return divisionid;
 	}
 }
