+++ src/main/java/com/github/timtebeek/taxonomy/model/Gencode.java
@@ -1,21 +1,22 @@
 package com.github.timtebeek.taxonomy.model;
 
-import lombok.Data;
-import lombok.EqualsAndHashCode;
 import org.neo4j.ogm.annotation.Index;
 import org.neo4j.ogm.annotation.NodeEntity;
+
+import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 @NodeEntity
 @Data
 @EqualsAndHashCode(callSuper = false, of = { "gencodeid" })
 public class Gencode extends AbstractEntity {
 	@Index(unique = true, primary = true)
-	Long	gencodeid;
+	Integer	gencodeid;
 
 	String	abbreviation, name, cde, starts;
 
 	@Override
-	public Long getId() {
+	public Integer getId() {
 		return gencodeid;
 	}
 }
